Move parts into components folder
We're going to eventually rewrite the tests to actually be
outside-in, creating new components the way the consumer
will, pushing data through those components the way the store
or DOM will, and checking (using chai-kefir) that the results are
what we expect. This will enable the tests to function more like
integration tests, so we can check that the system works
together the way we expect, rather than testing each
subcomponent separately.

Write tests, not too many, mostly integration.

This is the first step, treating `component` like its own package.
Eventually, `brookjs` itself will be a meta-package containing
everything you need to get started.
