+++ packages/brookjs/src/component/children/__tests__/util.js
