+++ packages/brookjs/src/component/children/__tests__/util.spec.js
