+++ packages/brookjs/src/component/children/child.js
