+++ packages/brookjs/src/component/children/util.js
