+++ packages/brookjs/src/component/helpers/animateAttribute.js
