+++ packages/brookjs/src/component/helpers/blackboxAttribute.js
