+++ packages/brookjs/src/component/helpers/containerAttribute.js
