+++ packages/brookjs/src/component/helpers/keyAttribute.js
