+++ packages/brookjs/src/component/helpers/mapActionTo.js
