+++ packages/brookjs/src/component/render/__tests__/fixtures/hideBlackboxed.hbs
