+++ packages/brookjs/src/component/render/__tests__/fixtures/simpleUpdate.hbs
