+++ packages/brookjs/src/component/render/__tests__/fixtures/updateChild.hbs
