+++ packages/brookjs/src/component/render/endAsObservable.js
@@ -1,4 +1,4 @@
-import Kefir from '../kefir';
+import Kefir from '../../kefir';
 
 /**
  * Create Transaction end stream.
