+++ packages/brookjs/src/component/render/getNodeKey.js
