+++ packages/brookjs/src/component/render/index.js
