+++ packages/brookjs/src/component/render/middleware.js
