+++ packages/brookjs/src/index.js
@@ -1,14 +1,10 @@
 import Kefir from './kefir';
-import children from './children';
 import combineActionReducers from './combineActionReducers';
-import component from './component';
+import { component, children, events, render, animateAttribute,
+    containerAttribute, blackboxAttribute, keyAttribute,
+    eventAttribute, mapActionTo, renderFromHTML, raf$ } from './component';
 import domDelta from './domDelta';
-import events from './events';
 import { observeDelta, ofType } from './observeDelta';
-import { animateAttribute, containerAttribute, blackboxAttribute,
-    keyAttribute, eventAttribute, mapActionTo } from './helpers';
-import render, { renderFromHTML } from './render';
-import { raf$ } from './rAF';
 
 export { Kefir, children, combineActionReducers, component, events, domDelta,
     observeDelta, animateAttribute, containerAttribute, eventAttribute, ofType,
